+++ extension.go
@@ -5,5 +5,5 @@ import "time"
 type Extension interface {
 	ServerListening(server Server)
 	IncomingRequest(req Request)
-	OutgoingResponse(req Request, res Response, resTime time.Duration, statusCode int16)
+	OutgoingResponse(req Request, res Response, resTime time.Duration, statusCode int32)
 }
